move getDustAppRunResultsFileTitle to run_dust_app
