+++ front/components/actions/dust_app_run/utils.ts
@@ -1,20 +1,4 @@
-import type { ModelId, SupportedFileContentType } from "@app/types";
-import { extensionsForContentType } from "@app/types/files";
-
-export function getDustAppRunResultsFileTitle({
-  appName,
-  resultsFileContentType,
-}: {
-  appName: string;
-  resultsFileContentType: SupportedFileContentType;
-}): string {
-  const extensions = extensionsForContentType(resultsFileContentType);
-  let title = `${appName}_output`;
-  if (extensions.length > 0) {
-    title += extensions[0];
-  }
-  return title;
-}
+import type { ModelId } from "@app/types";
 
 // TODO Daph refactor this we could simplify this.
 export type DustAppRunConfigurationType = {
